Add script and instructions to setup ~/hpc
Since local changes may not persist reboot and students may change PCs
on different days, ~/share is provided where work should be done.
However, ~/share doesn't support symlinks and the practicals and tools
like MAKER (used in the practical) rely heavily on symlinks. So we are
by passing ~/share and instead we are provide a script for students to
mount their HPC home directory to ~/hpc. All work must be done in ~/hpc.

Signed-off-by: Anurag Priyam <anurag.priyam@qmul.ac.uk>