Simplify read cleaning practical
Mask rare k-mers instead of trimming them. This retains read pairing and
lets us use cutadapt to remove short sequences while discarding unpaired
reads in the same step.

To deal with masked k-mers we use --trim-n option to remove leading and
trailing Ns in the reads. Masked k-mers in the middle of the reads are
left as it is.

Also, only mask singleton k-mers. Previously, we also eliminated k-mers
observed twice.

Signed-off-by: Anurag Priyam <anurag.priyam@qmul.ac.uk>