Standardised directory setup and add symlink commands (close #54)
I have added commands to symlink input files in the first three
practicals. The reason is that many students get it wrong in the first
two practicals resulting in failed commands. Most students get it after
the first two practicals so adding symlink commands to practicals 4 and
5 isn't absolutely necessary. Besides there is the tree output in
practicals 4 and 5 for the students to refer to. Finally, symlink
command in practical 2 introduces them to globbing which would be handy
in second part of practical 4 to symlink .bam and .bai files.

Signed-off-by: Anurag Priyam <anurag.priyam@qmul.ac.uk>